In general PluginFeatureFactories will just use the classes in their own package, so they can just access them without reflection. Move the complexity of componentFactory to the one factory which is truly dynamic: the property-based one
